pyproject.toml
[tool.poetry]
name = "course"
version = "0.1.0"
description = ""
authors = ["Rovshen Tagangylyjov <rovshentagangylyjov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
django-redis = {extras = ["hiredis"], version = "^5.4.0"}
django-filter = "^24.3"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
django-celery-results = "^2.5.1"
django-dbbackup = "^4.2.1"
django-auditlog = "^3.0.0"
django-modeltranslation = "^0.19.10"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-cachalot = "^2.7.0"
django-unfold = "^0.40.0"
pillow = "^11.0.0"
django-imagekit = "^5.0.0"
django-celery-email = "^3.0.0"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pylint-django = "^2.6.1"
flake8-isort = "^6.1.1"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.isort]
profile = 'black'
multi_line_output = 3
